Add PredicateWithSubquery as a base class for in and exists predicates
