--- h2/src/main/org/h2/expression/condition/ConditionExists.java
@@ -1,79 +1,0 @@
-/*
- * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
- * and the EPL 1.0 (https://h2database.com/html/license.html).
- * Initial Developer: H2 Group
- */
-package org.h2.expression.condition;
-
-import org.h2.command.dml.Query;
-import org.h2.engine.Session;
-import org.h2.expression.Expression;
-import org.h2.expression.ExpressionVisitor;
-import org.h2.result.ResultInterface;
-import org.h2.table.ColumnResolver;
-import org.h2.table.TableFilter;
-import org.h2.util.StringUtils;
-import org.h2.value.Value;
-import org.h2.value.ValueBoolean;
-
-/**
- * An 'exists' condition as in WHERE EXISTS(SELECT ...)
- */
-public class ConditionExists extends Condition {
-
-    private final Query query;
-
-    public ConditionExists(Query query) {
-        this.query = query;
-    }
-
-    @Override
-    public Value getValue(Session session) {
-        query.setSession(session);
-        ResultInterface result = query.query(1);
-        session.addTemporaryResult(result);
-        boolean r = result.hasNext();
-        return ValueBoolean.get(r);
-    }
-
-    @Override
-    public Expression optimize(Session session) {
-        session.optimizeQueryExpression(query);
-        return this;
-    }
-
-    @Override
-    public StringBuilder getSQL(StringBuilder builder, boolean alwaysQuote) {
-        builder.append("EXISTS(\n");
-        return StringUtils.indent(builder, query.getPlanSQL(alwaysQuote), 4, false).append(')');
-    }
-
-    @Override
-    public void updateAggregate(Session session, int stage) {
-        // TODO exists: is it allowed that the subquery contains aggregates?
-        // probably not
-        // select id from test group by id having exists (select * from test2
-        // where id=count(test.id))
-    }
-
-    @Override
-    public void mapColumns(ColumnResolver resolver, int level, int state) {
-        query.mapColumns(resolver, level + 1);
-    }
-
-    @Override
-    public void setEvaluatable(TableFilter tableFilter, boolean b) {
-        query.setEvaluatable(tableFilter, b);
-    }
-
-    @Override
-    public boolean isEverything(ExpressionVisitor visitor) {
-        return query.isEverything(visitor);
-    }
-
-    @Override
-    public int getCost() {
-        return query.getCostAsExpression();
-    }
-
-}
