+++ h2/src/main/org/h2/expression/condition/PredicateWithSubquery.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (http://h2database.com/html/license.html).
+ * Initial Developer: H2 Group
+ */
+package org.h2.expression.condition;
+
+import org.h2.command.dml.Query;
+import org.h2.engine.Session;
+import org.h2.expression.Expression;
+import org.h2.expression.ExpressionVisitor;
+import org.h2.table.ColumnResolver;
+import org.h2.table.TableFilter;
+import org.h2.util.StringUtils;
+
+/**
+ * Base class for predicates with a subquery.
+ */
+abstract class PredicateWithSubquery extends Condition {
+
+    final Query query;
+
+    PredicateWithSubquery(Query query) {
+        this.query = query;
+    }
+
+    @Override
+    public void mapColumns(ColumnResolver resolver, int level, int state) {
+        query.mapColumns(resolver, level + 1);
+    }
+
+    @Override
+    public Expression optimize(Session session) {
+        session.optimizeQueryExpression(query);
+        return this;
+    }
+
+    @Override
+    public void setEvaluatable(TableFilter tableFilter, boolean value) {
+        query.setEvaluatable(tableFilter, value);
+    }
+
+    @Override
+    public StringBuilder getSQL(StringBuilder builder, boolean alwaysQuote) {
+        return StringUtils.indent(builder.append('('), query.getPlanSQL(alwaysQuote), 4, false).append(')');
+    }
+
+    @Override
+    public void updateAggregate(Session session, int stage) {
+        query.updateAggregate(session, stage);
+    }
+
+    @Override
+    public boolean isEverything(ExpressionVisitor visitor) {
+        return query.isEverything(visitor);
+    }
+
+    @Override
+    public int getCost() {
+        return query.getCostAsExpression();
+    }
+
+}
